PIA26339: Checking the Cover on Perseverance's SHERLOC @nasa-jpl
Original Caption Released with Image:
Click here for animation (.gif, 16.7 MB)

Click here for animation (.mp4, 1.8 MB)

Video from a navigation camera aboard NASA's Perseverance Mars rover shows the position of a cover on the SHERLOC (Scanning Habitable Environments with Raman & Luminescence for Organics and Chemicals) instrument. During the video, the rover's robotic arm was commanded to move, allowing mission engineers to observe whether the cover for the Autofocus and Context Imager (ACI) camera would change position independent of the commanded motion. The imagery – acquired Jan. 23, 2024 (the 1,041 Martian day, or sol, of the mission) – indicated that the cover was not responsive.

On Jan. 6, 2024, a movable lens cover designed to protect the instrument's spectrometer and one of its cameras from dust became frozen in a position that prevented SHERLOC from collecting data. The rover team found a way to address the issue and confirmed the instrument is working on June 17, 2024.

A key objective for Perseverance's mission on Mars is astrobiology, including the search for signs of ancient microbial life. The rover is also characterizing the planet's geology and past climate, which paves the way for human exploration of the Red Planet. JPL, which is managed for NASA by Caltech in Pasadena, California, built and manages operations of the Perseverance rover.

For more about Perseverance: https://science.nasa.gov/mission/mars-2020-perseverance/

Image Credit:
NASA/JPL-Caltech

Image Addition Date:
2024-06-26