Sol 4664: Mars Hand Lens Imager (MAHLI) Nasa's Mars S
Using an onboard focusing process, the Mars Hand Lens Imager (MAHLI) aboard NASA's Mars rover Curiosity created this product by merging two to eight images previously taken by the MAHLI, located on the turret at the end of the rover's robotic arm.

Curiosity performed the merge on September 19, 2025, Sol 4664 of the Mars Science Laboratory Mission, at 07:36:30 UTC. The focus motor count position was 13942. This number indicates the lens position of the first image that was merged.

The onboard focus merge is sometimes performed on images acquired the same sol as the merge, and sometimes uses pictures obtained on an earlier sol. Focus merging is a method to make a composite of images of the same target acquired at different focus positions to bring all (or, as many as possible) features into focus in a single image. Because the MAHLI focus merge is performed on Mars, it also serves as a means to reduce the number of images sent back to Earth. Each focus merge produces two images: a color, best-focus product and a black-and-white image that scientists can use to estimate focus position for each element of the best focus product. Thus, up to eight images can be merged, reducing the number of images returned to Earth to two.

Credits: NASA/JPL-Caltech/MSSS  Curiosity Raw Images 
Sol 4664: Mars Hand Lens Imager (MAHLI)  Sol 4664: Mars Hand Lens Imager (MAHLI)
Nasa's Mars rover Curiosity acquired this image using its Mars Hand Lens Imager (MAHLI) on Sol 4664
Using an onboard focusing process, the Mars Hand Lens Imager (MAHLI) aboard NASA's Mars rover Curiosity created this product by merging two to eight images previously taken by the MAHLI, located on the turret at the end of the rover's robotic arm.

Curiosity performed the merge on September 19, 2025, Sol 4664 of the Mars Science Laboratory Mission, at 07:36:34 UTC. The focus motor count position was 13942. This number indicates the lens position of the first image that was merged.

The onboard focus merge is sometimes performed on images acquired the same sol as the merge, and sometimes uses pictures obtained on an earlier sol. Focus merging is a method to make a composite of images of the same target acquired at different focus positions to bring all (or, as many as possible) features into focus in a single image. Because the MAHLI focus merge is performed on Mars, it also serves as a means to reduce the number of images sent back to Earth. Each focus merge produces two images: a color, best-focus product and a black-and-white image that scientists can use to estimate focus position for each element of the best focus product. Thus, up to eight images can be merged, reducing the number of images returned to Earth to two.

Credits: NASA/JPL-Caltech/MSSS