feature image for Satellite.md
Anatomy of a Space Shuttle
![Shuttle](https://www.nasa.gov/wp-content/uploads/2025/07/0102619orig.jpg?resize=1024,467)
This illustration shows the parts of a space shuttle orbiter. About the same size and weight as a DC-9 aircraft, the orbiter contains the pressurized crew compartment (which can normally carry up to seven crew members), the cargo bay, and the three main engines mounted on its aft end.
NASA
This illustration labels important parts of a space shuttle orbiter. The orbiter was the heart and brains of the space shuttle and served as the crew transport vehicle that carried astronauts to and from space. The space shuttle was comprised of the orbiter, the main engines, the external tank, and the solid rocket boosters. The space shuttle was the world’s first reusable spacecraft and the first spacecraft in history that could carry large satellites both to and from orbit.

Image credit: @NASA [read](https://www.nasa.gov/image-article/anatomy-of-a-space-shuttle/) @dhs-gov